Bug fix for Question 8.11: integer variables overflow for the included test cases of n=100000 and n=300322. Use long instead of int.
The correct answer for n=100000 is 133423351001, not 279364825.
The correct answer for n=300322 is 3612344074348, not 276578412.

+++ Java/Ch 08. Recursion and Dynamic Programming/Q8_11_Coins/Tester.java
@@ -5,8 +5,8 @@ public class Tester {
 	public static void main(String[] args) {
 		for (int i = 0; i < 200; i++) {
 			int[] denoms = {25, 10, 5, 1};
-			int waysA = Question.makeChange(i, denoms);
-			int waysB = QuestionB.makeChange(i, denoms);
+			long waysA = Question.makeChange(i, denoms);
+			long waysB = QuestionB.makeChange(i, denoms);
 			if (waysA != waysB) {
 				System.out.println("Error: " + i + " : " + waysA + ", " + waysB);
 			}
